:recycle: determ record
