+++ src/record/document.rs
@@ -3,6 +3,23 @@ use crate::record::user::{RepresentUser, RequestUser};
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize, Clone, Serialize)]
+pub struct DetermRecord {
+    pub current: bool,
+    pub comment: String,
+    pub determ_date: String,
+    pub determ_titlename: String,
+    pub determ_usercd: String,
+    pub determ_username: String,
+    pub determ_status: String,
+    pub document_id: i32,
+    pub document_number: String,
+    pub phase_number: i32,
+    pub phase_title: String,
+    pub represent_username: String,
+    pub represent_usercd: String,
+}
+
+#[derive(Debug, Deserialize, Clone, Serialize)]
 pub struct SimulationDetermRecord {
     pub current: bool,
     pub comment: String,
