+++ src/response/document/document_determs.rs
@@ -1,3 +1,4 @@
+use crate::record::document::DetermRecord;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize, Clone, Serialize)]
@@ -14,21 +15,4 @@ pub struct GetDocumentDetermsResponseBody {
     pub total_count: i32,
     pub error: bool,
     pub records: Vec<DetermRecord>,
-}
-
-#[derive(Debug, Deserialize, Clone, Serialize)]
-pub struct DetermRecord {
-    pub current: bool,
-    pub comment: String,
-    pub determ_date: String,
-    pub determ_titlename: String,
-    pub determ_usercd: String,
-    pub determ_username: String,
-    pub determ_status: String,
-    pub document_id: i32,
-    pub document_number: String,
-    pub phase_number: i32,
-    pub phase_title: String,
-    pub represent_username: String,
-    pub represent_usercd: String,
 }
